+++ projects/plugins/jetpack/_inc/client/components/settings-card/index.jsx
@@ -13,7 +13,14 @@ import {
 	FEATURE_SPAM_AKISMET_PLUS,
 	FEATURE_SEARCH_JETPACK,
 	FEATURE_SIMPLE_PAYMENTS_JETPACK,
+	FEATURE_NEWSLETTER_JETPACK,
+	FEATURE_DOWNTIME_MONITORING_JETPACK,
+	FEATURE_SSO,
+	FEATURE_JETPACK_SOCIAL,
+	FEATURE_POST_BY_EMAIL,
 	getJetpackProductUpsellByFeature,
+	FEATURE_JETPACK_BLAZE,
+	FEATURE_JETPACK_EARN,
 } from 'lib/plans/constants';
 import { get, includes } from 'lodash';
 import ProStatus from 'pro-status';
@@ -22,19 +29,20 @@ import {
 	isSearchNewPricingLaunched202208,
 } from 'product-descriptions/utils';
 import PropTypes from 'prop-types';
-import React from 'react';
 import { connect } from 'react-redux';
 import { isAkismetKeyValid, isCheckingAkismetKey, getVaultPressData } from 'state/at-a-glance';
 import {
 	hasConnectedOwner as hasConnectedOwnerSelector,
 	isOfflineMode,
 	connectUser,
+	isUnavailableInOfflineMode,
 } from 'state/connection';
 import {
 	getSiteAdminUrl,
 	getUpgradeUrl,
 	isMultisite,
 	userCanManageModules,
+	shouldInitializeBlaze,
 } from 'state/initial-state';
 import { getModuleOverride, getModule } from 'state/modules';
 import { siteHasFeature, isFetchingSiteData } from 'state/site';
@@ -43,6 +51,7 @@ export const SettingsCard = inprops => {
 	const props = {
 		action: '',
 		saveDisabled: false,
+		isDisabled: false,
 		...inprops,
 	};
 
@@ -86,6 +95,7 @@ export const SettingsCard = inprops => {
 		return <span />;
 	}
 
+	const isDisabled = props.isDisabled;
 	const isSaving = props.saveDisabled,
 		feature = props.feature ? props.feature : false;
 	let header = props.header ? props.header : '';
@@ -223,6 +233,113 @@ export const SettingsCard = inprops => {
 					/>
 				);
 
+			case FEATURE_DOWNTIME_MONITORING_JETPACK:
+				if ( props.hasConnectedOwner || props.inOfflineMode ) {
+					return '';
+				}
+
+				return (
+					<JetpackBanner
+						title={ __(
+							'Connect your WordPress.com account to set up your status alerts.',
+							'jetpack'
+						) }
+						callToAction={ connectLabel() }
+						plan={ getJetpackProductUpsellByFeature( FEATURE_DOWNTIME_MONITORING_JETPACK ) }
+						feature={ feature }
+						onClick={ handleConnectClick( feature ) }
+						rna
+					/>
+				);
+
+			case FEATURE_SSO:
+				if ( props.hasConnectedOwner || props.inOfflineMode ) {
+					return '';
+				}
+
+				return (
+					<JetpackBanner
+						title={ __( 'Connect your WordPress.com account to enable Secure Sign-On', 'jetpack' ) }
+						callToAction={ connectLabel() }
+						plan={ getJetpackProductUpsellByFeature( FEATURE_SSO ) }
+						feature={ feature }
+						onClick={ handleConnectClick( feature ) }
+						rna
+					/>
+				);
+
+			case FEATURE_POST_BY_EMAIL:
+				if ( props.hasConnectedOwner || props.isUnavailableInOfflineMode( 'post-by-email' ) ) {
+					return '';
+				}
+
+				return (
+					<JetpackBanner
+						title={ __(
+							'Connect your WordPress.com account to enable publishing via email.',
+							'jetpack'
+						) }
+						callToAction={ connectLabel() }
+						plan={ getJetpackProductUpsellByFeature( FEATURE_POST_BY_EMAIL ) }
+						feature={ feature }
+						onClick={ handleConnectClick( feature ) }
+						rna
+					/>
+				);
+
+			case FEATURE_JETPACK_SOCIAL:
+				if ( props.hasConnectedOwner || props.inOfflineMode ) {
+					return '';
+				}
+
+				return (
+					<JetpackBanner
+						title={ __(
+							'Connect your WordPress.com account to add your social media accounts.',
+							'jetpack'
+						) }
+						callToAction={ connectLabel() }
+						plan={ getJetpackProductUpsellByFeature( FEATURE_JETPACK_SOCIAL ) }
+						feature={ feature }
+						onClick={ handleConnectClick( feature ) }
+						rna
+					/>
+				);
+
+			case FEATURE_JETPACK_BLAZE:
+				if ( props.blazeAvailable.can_init || props.inOfflineMode ) {
+					return '';
+				}
+
+				return (
+					<JetpackBanner
+						title={ __(
+							'Connect your WordPress.com account to set up campaigns and promote your content.',
+							'jetpack'
+						) }
+						callToAction={ connectLabel() }
+						plan={ getJetpackProductUpsellByFeature( FEATURE_JETPACK_BLAZE ) }
+						feature={ feature }
+						onClick={ handleConnectClick( feature ) }
+						rna
+					/>
+				);
+
+			case FEATURE_JETPACK_EARN:
+				return (
+					<JetpackBanner
+						title={ __(
+							'Connect your WordPress.com account to discover tools to earn money with your site.',
+							'jetpack'
+						) }
+						callToAction={ connectLabel() }
+						plan={ getJetpackProductUpsellByFeature( FEATURE_JETPACK_EARN ) }
+						feature={ feature }
+						onClick={ handleConnectClick( feature ) }
+						rna
+					/>
+				);
+
 			case FEATURE_GOOGLE_ANALYTICS_JETPACK:
 				if ( props.hasGoogleAnalytics ) {
 					return '';
@@ -322,6 +439,22 @@ export const SettingsCard = inprops => {
 						rna
 					/>
 				);
+
+			case FEATURE_NEWSLETTER_JETPACK:
+				if ( props.hasConnectedOwner ) {
+					return '';
+				}
+
+				return (
+					<JetpackBanner
+						title={ __( 'Connect your WordPress.com account to enable newsletters.', 'jetpack' ) }
+						callToAction={ connectLabel() }
+						plan={ getJetpackProductUpsellByFeature( FEATURE_NEWSLETTER_JETPACK ) }
+						feature={ feature }
+						onClick={ handleConnectClick( feature ) }
+						rna
+					/>
+				);
 			default:
 				return '';
 		}
@@ -414,11 +547,17 @@ export const SettingsCard = inprops => {
 			<form
 				{ ...( moduleId ? { id: moduleId } : null ) }
 				className={ `jp-form-settings-card` }
-				onSubmit={ ! isSaving ? props.onSubmit : undefined }
+				onSubmit={ ! isDisabled && ! isSaving ? props.onSubmit : undefined }
 			>
 				<SectionHeader label={ header }>
 					{ ! props.hideButton && (
-						<Button primary rna compact type="submit" disabled={ isSaving || ! props.isDirty() }>
+						<Button
+							primary
+							rna
+							compact
+							type="submit"
+							disabled={ isDisabled || isSaving || ! props.isDirty() }
+						>
 							{ isSaving
 								? _x( 'Saving…', 'Button caption', 'jetpack' )
 								: _x(
@@ -447,6 +586,7 @@ export const SettingsCard = inprops => {
 SettingsCard.propTypes = {
 	action: PropTypes.string,
 	saveDisabled: PropTypes.bool,
+	isDisabled: PropTypes.bool,
 };
 
 export default connect(
@@ -478,6 +618,8 @@ export default connect(
 			hasSimplePayments: siteHasFeature( state, 'simple-payments' ),
 			hasVideoPress: siteHasFeature( state, 'videopress' ),
 			hasWordAds: siteHasFeature( state, 'wordads' ),
+			isUnavailableInOfflineMode: module_name => isUnavailableInOfflineMode( state, module_name ),
+			blazeAvailable: shouldInitializeBlaze( state ),
 		};
 	},
 	dispatch => ( {