+++ projects/plugins/jetpack/extensions/extended-blocks/core-site-logo/index.tsx
@@ -26,6 +26,7 @@ type CoreSelect = {
 		url: string;
 		title: string;
 		description: string;
+		site_logo: number;
 	};
 };
 
@@ -80,6 +81,7 @@ const useSiteDetails = () => {
 		domain: window?.Jetpack_Editor_Initial_State?.siteFragment,
 		name: siteSettings?.title,
 		description: siteSettings?.description,
+		siteLogo: siteSettings?.site_logo || 0,
 	};
 };
 
@@ -123,13 +125,11 @@ const siteLogoEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 
 		const siteDetails = useSiteDetails();
 
-		const styleLogoFeatureEnabled = getFeatureAvailability( 'ai-logo-style-selector-support' );
-
 		return (
 			<>
 				<BlockEdit { ...props } />
 				<BlockControls group="block">
-					<AiToolbarButton clickHandler={ showModal } />
+					<AiToolbarButton showButtonText={ ! siteDetails?.siteLogo } clickHandler={ showModal } />
 				</BlockControls>
 				<GeneratorModal
 					isOpen={ isLogoGeneratorModalVisible }
@@ -139,7 +139,6 @@ const siteLogoEditWithAiComponents = createHigherOrderComponent( BlockEdit => {
 					context={ PLACEMENT_CONTEXT }
 					placement={ TOOL_PLACEMENT }
 					siteDetails={ siteDetails }
-					showStyleSelector={ styleLogoFeatureEnabled }
 				/>
 			</>
 		);