+++ projects/packages/videopress/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.24.11] - 2024-10-07
+### Changed
+- Update connection nudge for VideoPress connection banner in blocks. [#39628]
+- Updated package dependencies. [#39594]
+
+### Fixed
+- VideoPress: fix thumbnail update bug that showed the old thumbnail for a couple seconds on the details page. [#39631]
+
+## [0.24.10] - 2024-09-30
+### Changed
+- Update dependencies. [#39528]
+
 ## [0.24.9] - 2024-09-23
 ### Changed
 - Update dependencies. [#38595]
@@ -1444,6 +1456,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Created empty package [#24952]
 
+[0.24.11]: https://github.com/Automattic/jetpack-videopress/compare/v0.24.10...v0.24.11
+[0.24.10]: https://github.com/Automattic/jetpack-videopress/compare/v0.24.9...v0.24.10
 [0.24.9]: https://github.com/Automattic/jetpack-videopress/compare/v0.24.8...v0.24.9
 [0.24.8]: https://github.com/Automattic/jetpack-videopress/compare/v0.24.7...v0.24.8
 [0.24.7]: https://github.com/Automattic/jetpack-videopress/compare/v0.24.6...v0.24.7